moved raise and log function to setup logging and moved setup logging file to source folder
