+++ src/setup_logging.py
@@ -8,3 +8,8 @@ formatter = logging.Formatter(
 handler = logging.StreamHandler()
 handler.setFormatter(formatter)
 logger.addHandler(handler)
+
+
+def raise_and_log(logged_exception: Exception, exception_message: str):
+    logger.error(exception_message)
+    raise logged_exception(exception_message)
